% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{umap_defaults}
\alias{umap_defaults}
\title{UMAP defaults}
\usage{
umap_defaults()
}
\value{
Named list of parameters with their defaults.
}
\description{
Default parameters for the UMAP algorithm.
Values here are passed to the \pkg{umappp} C++ library.
}
\examples{
defaults()

}
\author{
Aaron Lun
}
