% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_from_matrix.R
\name{umap_from_matrix}
\alias{umap_from_matrix}
\title{UMAP from an input matrix}
\usage{
umap_from_matrix(
  y,
  ...,
  method = c("Annoy", "VPTree"),
  ndim = 2,
  nthreads = 1,
  tick = 0
)
}
\arguments{
\item{y}{Numeric matrix of observations (rows) and variables (columns).}

\item{...}{Further arguments to change the defaults, see \code{\link{umap_defaults}}.}

\item{method}{String specifying the neighbor detection algorithm to use.}

\item{ndim}{Integer scalar specifying the number of embedding dimensions to return.}

\item{nthreads}{Integer scalar specifying the number of threads to use.}

\item{tick}{Integer scalar specifying the epoch interval to return embeddings in progress.}
}
\value{
If \code{tick=0}, a single matrix containing the embedding.
This has \code{ndim} columns and number of rows equal to \code{nrow(y)}.

If \code{tick > 0}, a list of such matrices containing the embedding at each tick.
}
\description{
Compute UMAP coordinates from an input matrix of observation-level data.
This involves a round of nearest neighbor detection before the UMAP iterations themselves.
}
\examples{
y <- as.matrix(iris[,1:4])
out <- umap_from_matrix(y)
plot(out[,1], out[,2], col=iris[,5])

}
\author{
Aaron Lun
}
